requirements.txt
python-dotenv
fastapi
uvicorn
pydantic
langchain
langchain-google-genai
